Move horarios para itinerario em vez de segmento
