--- app/src/main/java/com/thoughtworks/aceleradora/airline/dominio/AeroportoRepository.java
@@ -1,6 +1,0 @@
-package com.thoughtworks.aceleradora.airline.dominio;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface AeroportoRepository extends CrudRepository<Aeroporto, Long> {
-}
