+++ app/src/main/java/com/thoughtworks/aceleradora/airline/dominio/Itinerario.java
@@ -2,6 +2,8 @@ package com.thoughtworks.aceleradora.airline.dominio;
 
 import javax.persistence.*;
 
+import java.time.LocalDateTime;
+
 import static javax.persistence.CascadeType.*;
 import static javax.persistence.GenerationType.IDENTITY;
 
@@ -20,6 +22,12 @@ public class Itinerario {
     @JoinColumn(name = "destino_aeroporto_id")
     private Aeroporto destino;
 
+    @Column(name = "horario_decolagem")
+    private LocalDateTime horarioDecolagem;
+
+    @Column(name = "horario_pouso")
+    private LocalDateTime horarioPouso;
+
     public Itinerario() {
     }
 
@@ -37,5 +45,21 @@ public class Itinerario {
 
     public void setDestino(Aeroporto destino) {
         this.destino = destino;
+    }
+
+    public LocalDateTime getHorarioDecolagem() {
+        return horarioDecolagem;
+    }
+
+    public void setHorarioDecolagem(LocalDateTime horarioDecolagem) {
+        this.horarioDecolagem = horarioDecolagem;
+    }
+
+    public LocalDateTime getHorarioPouso() {
+        return horarioPouso;
+    }
+
+    public void setHorarioPouso(LocalDateTime horarioPouso) {
+        this.horarioPouso = horarioPouso;
     }
 }
