+++ app/src/main/resources/db/migration/V5__Alter_itinerario_table.sql
@@ -1,0 +1,2 @@
+ALTER TABLE itinerario ADD horario_decolagem TIMESTAMP;
+ALTER TABLE itinerario ADD horario_pouso TIMESTAMP;
