+++ app/src/main/resources/db/migration/V6__Alter_segmento_table.sql
@@ -1,0 +1,2 @@
+ALTER TABLE segmento DROP COLUMN horario_decolagem;
+ALTER TABLE segmento DROP COLUMN horario_pouso;
